+++ src/parser/value_expr.rs
@@ -1,18 +1,14 @@
+use super::utils::parse_rational;
+use super::{GrammarParser, Rule};
 use crate::app;
 use crate::models::{Account, Currency, Money, Payee, Posting, Transaction};
-use crate::pest::Parser;
 use crate::List;
 use chrono::NaiveDate;
-use num::{abs, BigInt, BigRational};
+use num::{abs, BigRational};
+use pest::Parser;
 use regex::Regex;
-use std::rc::Rc;
-use std::str::FromStr;
-
 use std::collections::HashMap;
-
-#[derive(Parser)]
-#[grammar = "grammar/expressions.pest"]
-pub struct ValueExpressionParser;
+use std::rc::Rc;
 
 pub fn eval_expression(
     expression: &str,
@@ -21,7 +17,7 @@ pub fn eval_expression(
     commodities: &mut List<Currency>,
     regexes: &mut HashMap<String, Regex>,
 ) -> EvalResult {
-    let parsed = ValueExpressionParser::parse(Rule::value_expr, expression)
+    let parsed = GrammarParser::parse(Rule::value_expr, expression)
         .expect("unsuccessful parse") // unwrap the parse result
         .next()
         .unwrap()
@@ -174,9 +170,9 @@ pub fn eval(
             let right = eval(rhs, posting, transaction, commodities, regexes);
             match op {
                 Binary::Eq => {
-                    // println!("{:?} eq {:?}", left, right); // todo delete
                     match right {
                         EvalResult::Regex(rhs) => match left {
+                            // TODO regex comparison with accounts is one source of slow speed
                             EvalResult::Account(lhs) => EvalResult::Boolean(lhs.is_match(rhs)),
                             EvalResult::Payee(lhs) => EvalResult::Boolean(lhs.is_match(rhs)),
                             EvalResult::String(lhs) => match lhs {
@@ -280,15 +276,27 @@ pub fn eval(
                 }
                 Binary::Or | Binary::And => {
                     if let EvalResult::Boolean(lhs) = left {
-                        if let EvalResult::Boolean(rhs) = right {
-                            EvalResult::Boolean(match op {
-                                Binary::Or => lhs | rhs,
-                                Binary::And => lhs & rhs,
-                                _ => unreachable!(),
-                            })
-                        } else {
-                            panic!("Should be booleans")
-                        }
+                        EvalResult::Boolean(match op {
+                            Binary::Or => {
+                                if lhs {
+                                    true
+                                } else if let EvalResult::Boolean(rhs) = right {
+                                    rhs
+                                } else {
+                                    panic!("Should be booleans")
+                                }
+                            }
+                            Binary::And => {
+                                if !lhs {
+                                    false
+                                } else if let EvalResult::Boolean(rhs) = right {
+                                    rhs
+                                } else {
+                                    panic!("Should be booleans")
+                                }
+                            }
+                            _ => unreachable!(),
+                        })
                     } else {
                         panic!("Should be booleans")
                     }
@@ -393,16 +401,16 @@ fn build_ast_from_expr(
                     match child.as_rule() {
                         Rule::number => Node::Money {
                             currency: money.next().unwrap().as_str().to_string(),
-                            amount: parse_big_rational(child.as_str()),
+                            amount: parse_rational(child),
                         },
                         Rule::currency => Node::Money {
                             currency: child.as_str().to_string(),
-                            amount: parse_big_rational(money.next().unwrap().as_str()),
+                            amount: parse_rational(money.next().unwrap()),
                         },
                         unknown => panic!("Unknown rule: {:?}", unknown),
                     }
                 }
-                Rule::number => Node::Number(parse_big_rational(first.as_str())),
+                Rule::number => Node::Number(parse_rational(first)),
                 Rule::regex | Rule::string => {
                     let full = first.as_str().to_string();
                     let n = full.len() - 1;
@@ -450,24 +458,4 @@ fn parse_unary_expr(operation: Unary, child: Node) -> Node {
         op: operation,
         child: Box::new(child),
     }
-}
-
-fn parse_big_rational(input: &str) -> BigRational {
-    let mut num = String::new();
-    let mut den = "1".to_string();
-    let mut decimal = false;
-    for c in input.chars() {
-        if c == '.' {
-            decimal = true
-        } else {
-            num.push(c);
-            if decimal {
-                den.push('0')
-            };
-        }
-    }
-    BigRational::new(
-        BigInt::from_str(num.as_str()).unwrap(),
-        BigInt::from_str(den.as_str()).unwrap(),
-    )
 }