+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 
+## [0.13.0] - 2021-02-27 
+### Added
+- Improved documentation
+- Support for [hledger syntax for payees](https://github.com/frosklis/dinero-rs/issues/37)
+### Fixed
+- keep tags from transactions
+- match automated transactions only once per transaction, like ```ledger``` does
+- enable comments in price ```p``` directives
 ## [0.12.0] - 2021-02-24
 ### Added
 - support for (some of the) automated transaction syntax, what Claudio uses in his personal ledger
 ### Fixed
 - speed bump (from 44 seconds to 7 seconds) in a big personal ledger
-- 
+
 ## [0.11.1] - 2021-02-22
 ### Fixed
 - Fixed bug in balance report